Bug-fix: Functions assigned to input props are not scoped properly
Uses a proxy object to store the definition scope alongside the function.
This scope can be updated during the lifetime of the function, serving as
a poor-man's version of multiple `bind` calls to it.  Additionally, the
function scope is combined with the execution scope in order to support
use-cases where the invocation scope is the only way to access certain
arguments.